Add functionality to delete item from list
